Remove blue line and add shadow to Input Field
