+++ duckchat/duckchat-impl/src/main/java/com/duckduckgo/duckchat/impl/inputscreen/ui/InputScreenFragment.kt
@@ -220,7 +220,6 @@ class InputScreenFragment : DuckDuckGoFragment(R.layout.fragment_input_screen) {
 
     private fun exitInterstitial() {
         hideKeyboard(binding.inputModeWidget.inputField)
-        binding.inputModeWidget.animateOmnibarFocusedState(false)
         requireActivity().supportFinishAfterTransition()
     }
