+++ duckchat/duckchat-impl/src/main/res/layout/fragment_input_screen.xml
@@ -19,7 +19,8 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/root"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:clipChildren="false">
 
     <com.duckduckgo.duckchat.impl.inputscreen.ui.view.InputModeWidget
         android:id="@+id/inputModeWidget"
